Ensure columns exist after spread()
tidyr::spread() promotes labels observed in the dataset into columns.
Sometimes labels that should "always" be present will be missing because
we don't have enough clients in a particular branch or something.
`ensure()` is a helper function that tries to guarantee that a column
exists after labels have been promoted, by filling with NA.